+++ gpxviewr/baseweb/admin.py
@@ -3,6 +3,7 @@ from django.contrib import admin
 from baseweb.models import (
     GPXTrack,
     GPXTrackWayPoint,
+    GPXTrackSegmentPoint,
 )
 
 
@@ -14,5 +15,10 @@ class GPXTrackWayPointAdmin(admin.ModelAdmin):
     pass
 
 
+class GPXTrackSegmentPointAdmin(admin.ModelAdmin):
+    pass
+
+
 admin.site.register(GPXTrack, GPXTrackAdmin)
 admin.site.register(GPXTrackWayPoint, GPXTrackWayPointAdmin)
+admin.site.register(GPXTrackSegmentPoint, GPXTrackSegmentPointAdmin)