+++ gpxviewr/baseweb/forms.py
@@ -5,7 +5,7 @@ from django.core.files.base import File
 from django.db.models.base import Model
 from django.forms.utils import ErrorList
 
-from .models import GPXTrack, GPXWayPointType, generate_default_delete_after_date
+from .models import GPXFile, GPXWayPointType, generate_default_delete_after_date
 
 
 def waypoint_types_choices():
@@ -16,13 +16,13 @@ def waypoint_types_choices():
     return c
 
 
-class GPXTrackUploadForm(forms.ModelForm):
+class GPXFileUploadForm(forms.ModelForm):
     wpt_options = forms.JSONField()
     file = forms.FileField()
     delete_after = forms.DateField(initial=generate_default_delete_after_date)
 
     class Meta:
-        model = GPXTrack
+        model = GPXFile
         fields = ["file", "wpt_options", "delete_after"]